skills/xiaohongshu-pub/scripts/requirements.txt
# 小红书HTML发布工具依赖清单
# Xiaohongshu HTML Publisher Requirements

# HTML解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 浏览器自动化
selenium>=4.15.0

# 图片处理
Pillow>=10.0.0

# AI智能（可选，用于合集分类）
zhipuai>=0.1.0

# 进程检查（可选）
# psutil>=5.9.0
